Poll 27 text mining
"../input/polling/Polling 27 - 22 - 29 Apr.csv" is the raw result of poll 27, which consists of 2 questions. One of the questions, "Bagikan cerita menarik Mums ketika Mums dinyatakan hamil (Max:105 karakter)" is a free-response text question. The purpose of this code is to clean the text responses and create a frequency table of unigrams and bigrams.